requirements.txt
atoma==0.0.17
beautifulsoup4==4.13.3
click==8.1.8
duckdb==1.2.1
flask==3.1.0
jinja2==3.1.6
pandas==2.2.3
pyarrow==19.0.1
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
tqdm==4.67.1
psycopg2==2.9.9  # PostgreSQL adapter
python-dotenv==1.0.1  # For loading environment variables
